prototype/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
fire==0.6.0
pandas==2.2.3
torch
tqdm
transformers==4.44.2
huggingface_hub[cli]==0.25.0
pytest==8.3.3
accelerate==0.34.2
